+++ src/lag.c
@@ -6,8 +6,8 @@
 #include <dsp/utils.h>
 
 static inline void update_(lag* self) {
-    self->t_sec += 0.0001;  // guard div/0
-    self->alpha_ = 1.0 - exp((-1.0 / (self->t_sec * self->sr)));
+    self->t_sec += 0.0001f;  // guard div/0
+    self->alpha_ = 1.0f - expf((-1.0 / (self->t_sec * self->sr)));
 }
 
 static inline float tick_(lag* self, float xn) {
@@ -18,8 +18,8 @@ static inline float tick_(lag* self, float xn) {
 void lag_init(lag* self, float t_sec, float sr) {
     self->sr = sr;
     self->t_sec = t_sec;
-    self->last_ = 0.0;
-    self->alpha_ = 0.0;
+    self->last_ = 0.0f;
+    self->alpha_ = 0.0f;
     update_(self);
 }
